requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
Pillow==7.1.2
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
